Move tests into __test__ folder
