+++ app/javascript/react/__tests__/components/Button.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, fireEvent } from '@testing-library/react';
-import Button from './Button';
+import Button from '../../components/Button';
 
 describe('Button Component', () => {
     // Snapshot tests
