+++ app/javascript/react/__tests__/components/Card.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import Card from './Card';
+import Card from '../../components/Card';
 
 describe('Card Component', () => {
     // Snapshot tests
