+++ app/javascript/react/__tests__/components/LoadingSpinner.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import LoadingSpinner from './LoadingSpinner';
+import LoadingSpinner from '../../components/LoadingSpinner';
 
 describe('LoadingSpinner Component', () => {
     // Snapshot tests
